+++ MegaManLofi/ConsoleRenderDefs.h
@@ -20,8 +20,8 @@ namespace MegaManLofi
       short ArenaViewportLeftChars = 0;
       short ArenaViewportTopChars = 0;
 
-      long long ArenaCharWidth = 0;
-      long long ArenaCharHeight = 0;
+      float ArenaCharWidth = 0;
+      float ArenaCharHeight = 0;
 
       short ArenaViewportWidthChars = 0;
       short ArenaViewportHeightChars = 0;
@@ -30,10 +30,10 @@ namespace MegaManLofi
       short ArenaStatusBarTopChars = 0;
       short ArenaStatusBarWidthChars = 0;
 
-      double PitfallAnimationSeconds = 0;
+      float PitfallAnimationSeconds = 0;
 
-      double PlayerExplosionAnimationSeconds = 0;      
-      long long PlayerExplosionParticleVelocity = 0;
+      float PlayerExplosionAnimationSeconds = 0;      
+      float PlayerExplosionParticleVelocity = 0;
 
       ConsoleColor DefaultForegroundColor = (ConsoleColor)0;
       ConsoleColor DefaultBackgroundColor = (ConsoleColor)0;
@@ -60,8 +60,8 @@ namespace MegaManLofi
 
       ConsoleImage PlayingMenuPlayerImage;
 
-      long long PlayerThwipVelocity = 0;
-      double PlayerPostThwipDelaySeconds = 0;
+      float PlayerThwipVelocity = 0;
+      float PlayerPostThwipDelaySeconds = 0;
 
       short TitleTextLeftChars = 0;
       short TitleTextTopChars = 0;
@@ -76,10 +76,10 @@ namespace MegaManLofi
       short TitleKeyBindingsMiddleXChars = 0;
       short TitleKeyBindingsTopChars = 0;
       int TitleStarCount = 0;
-      long long MinTitleStarVelocity = 0;
-      long long MaxTitleStarVelocity = 0;
+      float MinTitleStarVelocity = 0;
+      float MaxTitleStarVelocity = 0;
 
-      double GetReadyAnimationSeconds = 0;
+      float GetReadyAnimationSeconds = 0;
 
       ConsoleImage PauseOverlayImage;
       ConsoleImage GameOverImage;
