+++ MegaManLofiTests/mock_FrameRateProvider.h
@@ -10,7 +10,7 @@ public:
    MOCK_METHOD( long long, GetMinimumFrameRate, ( ), ( const, override ) );
    MOCK_METHOD( bool, HasMinimumFrameRate, ( ), ( const, override ) );
    MOCK_METHOD( long long, GetCurrentFrame, ( ), ( const, override ) );
-   MOCK_METHOD( double, GetFrameSeconds, ( ), ( const, override ) );
+   MOCK_METHOD( float, GetFrameSeconds, ( ), ( const, override ) );
    MOCK_METHOD( long long, GetLagFrameCount, ( ), ( const, override ) );
    MOCK_METHOD( long long, GetElapsedNanoseconds, ( ), ( const, override ) );
    MOCK_METHOD( long long, GetAverageFrameRate, ( ), ( const, override ) );
